feat: add backend intergration for datatables, invoices

+++ tests/api/routes/test_users.py
@@ -99,9 +99,11 @@ def test_get_existing_user_current_user(client: TestClient, db: Session) -> None
         "password": password,
     }
     r = client.post(f"{settings.API_V1_STR}/login/access-token", data=login_data)
-    tokens = r.json()
-    a_token = tokens["access_token"]
-    headers = {"Authorization": f"Bearer {a_token}"}
+    r.raise_for_status()
+    csrf_token = client.cookies.get("csrf_token")
+    if not csrf_token:
+        raise KeyError("csrf_token")
+    headers = {"X-CSRF-Token": csrf_token}
 
     r = client.get(
         f"{settings.API_V1_STR}/users/{user_id}",
@@ -429,9 +431,11 @@ def test_delete_user_me(client: TestClient, db: Session) -> None:
         "password": password,
     }
     r = client.post(f"{settings.API_V1_STR}/login/access-token", data=login_data)
-    tokens = r.json()
-    a_token = tokens["access_token"]
-    headers = {"Authorization": f"Bearer {a_token}"}
+    r.raise_for_status()
+    csrf_token = client.cookies.get("csrf_token")
+    if not csrf_token:
+        raise KeyError("csrf_token")
+    headers = {"X-CSRF-Token": csrf_token}
 
     r = client.delete(
         f"{settings.API_V1_STR}/users/me",
